fix(agenda_f3): ETL atômico + matrícula BIGINT + typo Altamuro→Altamiro
3 correções pós-deploy:

1. ETL: load_dotenv tinha que ser ANTES do import de f3_db
   f3_db cria o engine do Postgres em tempo de import (precisa de
   PG_RDS_HOST). O load_dotenv estava dentro de run() — quebrava com
   KeyError: 'PG_RDS_HOST' antes mesmo de chegar em run().

2. Schema: matricula, dias_agend_cons, idendereco → BIGINT
   psycopg2.errors.NumericValueOutOfRange em 10/13 postos no primeiro
   run. INTEGER comporta até ~2.1bi mas algum campo do CAMIM passou
   disso (provavelmente Dif_dias_agend_cons num caso extremo). Bumpar
   pra BIGINT custa zero e elimina a classe de bug.

3. UI (KPI + F3): "Altamuro" → "Altamiro" no optgroup do dropdown
   Typo no nome do dono do grupo — corrigido nos DOIS arquivos
   agenda_dia.html (KPI raiz e F3) já que o template do F3 foi cópia.

Estado pós-fixes:
- ETL rodou 13/13 postos OK, 4912 pacientes gravados em f3.agenda_dia
- camila2.ia.camim.com.br/agenda_dia respondendo (302 → IDCamim login)
- Banco RDS f3 com snapshot do dia + meta de cada posto
- Cron */30 6-22 ativo

+++ agenda_f3/f3_db.py
@@ -45,7 +45,7 @@ class AgendaDia(Base):
     id               = Column(BigInteger, primary_key=True, autoincrement=True)
     posto            = Column(String(1), nullable=False)
     data             = Column(Date,      nullable=False)
-    matricula        = Column(Integer)
+    matricula        = Column(BigInteger)        # bigint p/ segurança
     cfcliente        = Column(String(4))
     posto_cliente    = Column(String(4))
     paciente         = Column(Text)
@@ -54,12 +54,12 @@ class AgendaDia(Base):
     medico           = Column(Text)
     hora_prevista    = Column(String(5))
     hora_confirmacao = Column(String(5))
-    dias_agend_cons  = Column(Integer)
+    dias_agend_cons  = Column(BigInteger)        # bigint p/ casos extremos DATEDIFF
     atendido         = Column(Text)
     desistencia      = Column(SmallInteger, default=0, nullable=False)
     situacao         = Column(Text)
     pagou_no_dia     = Column(Boolean, default=False, nullable=False)
-    idendereco       = Column(Integer)
+    idendereco       = Column(BigInteger)
     gerado_em        = Column(DateTime(timezone=True), nullable=False)
 
     __table_args__ = (
